Added submitOnPause feature
submitOnPause is an autocomplete feature. When enabled, if you type in
an address that causes the autocomplete to come up with no results,
after 5 seconds the plugin will submit the address as freeform to the
street api. If it comes back valid, it will be placed in the suggestion
container (1 lookup used). When click on it will populate the form and
accept the address without using an additional lookup.